+++ DomainLogic/Model/DepositTerm.cs
@@ -4,9 +4,9 @@ namespace DomainLogic.Model
 {
     public class DepositTerm
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
-        public byte CurrencyId { get; set; }
+        public int CurrencyId { get; set; }
 
         [Display(Name = "Months")]
         public byte Months { get; set; }
